requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
astor==0.8.1
astpretty==3.0.0
attrs==23.2.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
execnet==2.1.1
filelock==3.15.4
fonttools==4.53.1
future-fstrings==1.2.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
networkx==3.3
numpy==2.0.1
outcome==1.3.0.post0
packaging==24.1
page==0.2
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
playwright==1.45.1
pluggy==1.5.0
pretty==0.1
pycparser==2.22
pyee==11.1.0
Pygments==2.18.0
pyodbc==5.1.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.2
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-order==1.3.0
pytest-playwright==0.5.1
pytest-pretty==1.2.0
pytest-pretty-terminal==1.1.0
pytest-pspec==0.0.4
pytest-rerun==0.0.1
pytest-rerunfailures==14.0
pytest-sugar==1.0.0
pytest-testdox==3.1.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
rerun==1.0.30
rich==13.7.1
screeninfo==0.8.1
selenium==4.26.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
termcolor==2.4.0
text-unidecode==1.3
tqdm==4.66.5
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
websocket-client==1.8.0
wsproto==1.2.0
Faker==35.2.0